fix: comprehensive AMP dtype consistency in EGNN and Denoiser
- EGNNLayer: ensure all tensors created for aggregation (degree, coord_update,
  m_i) match the working dtype under AMP (float16 for h, float32 for x)
- Convert dist_sq to work_dtype before torch.cat to avoid mixed-dtype
  promotion and keep AMP speed benefit
- Keep coordinate update path in float32 for molecular geometry precision
- Cast coord_head output to float32 in Denoiser for coordinate loss stability
- Fix EGNNEmbedding time/condition expansion: use .expand() instead of
  .unsqueeze(0).expand() which broke for both single and batched molecules

+++ symmcd_diffusion/models/denoiser.py
@@ -135,8 +135,8 @@ class Denoiser(nn.Module):
                 h, positions, edge_index, edge_attr, None, node_mask
             )
 
-        # Predict coordinate noise
-        coord_noise = self.coord_head(h)
+        # Predict coordinate noise (keep float32 for coordinate precision)
+        coord_noise = self.coord_head(h).float()
 
         # Predict clean atom types
         atom_logits = self.atom_head(h)
@@ -150,6 +150,6 @@ class Denoiser(nn.Module):
             "coord_noise": coord_noise,
             "atom_logits": atom_logits,
             "bond_logits": bond_logits,
-            "h": h,           # final node features (for auxiliary losses)
-            "positions": positions_out,  # final positions (for debugging)
+            "h": h,
+            "positions": positions_out,
         }
